+++ internal/models/models.go
@@ -140,7 +140,8 @@ type CPUOptions struct {
 	HVIPI                  bool   `json:"hv_ipi" yaml:"hv_ipi"`
 	HVAVIC                 bool   `json:"hv_avic" yaml:"hv_avic"`
 	TopoExt                bool   `json:"topoext" yaml:"topoext"`
-	L3Cache                bool   `json:"l3_cache" yaml:"l3_cache"`
+	L3Cache                bool              `json:"l3_cache" yaml:"l3_cache"`
+	L3CacheSizeDie        map[int]string    `json:"l3_cache_size_die" yaml:"l3_cache_size_die"`
 	X2APIC                 bool   `json:"x2apic" yaml:"x2apic"`
 	Migratable             bool   `json:"migratable" yaml:"migratable"`
 	InvTSC                 bool   `json:"invtsc" yaml:"invtsc"`